Database work
Added a Gradle dependency for CSV reading - if it doesn't automatically install when you launch, I'm pretty sure you should be able to go to Build -> Make Project (on macOS, if on Windows I'm not sure).
More info on specifics in the Google Doc. Tested as thoroughly as I could but some bugs are bound to slip through, apologies in advance.